Move MiniSet to data_structures
remove the need for T to be copy from MiniSet as was done for MiniMap
